Rename exception name and move exception in package exception.
