+++ src/com/archosResearch/jCHEKS/engine/model/AbstractModel.java
@@ -1,5 +1,6 @@
 package com.archosResearch.jCHEKS.engine.model;
 
+import com.archosResearch.jCHEKS.engine.model.contact.exception.ContactAlreadyExistException;
 import com.archosResearch.jCHEKS.concept.engine.ModelObservable;
 import com.archosResearch.jCHEKS.engine.model.contact.Contact;
 import java.util.ArrayList;
@@ -10,7 +11,7 @@ import java.util.ArrayList;
  */
 public abstract class AbstractModel extends ModelObservable {
 
-    public abstract void addContact(Contact contact) throws NameOfContactAlreadyExistInContactsException;
+    public abstract void addContact(Contact contact) throws ContactAlreadyExistException;
 
     public abstract void addIncomingMessage(String messageContent, Contact contact);
