--- src/com/archosResearch/jCHEKS/engine/model/NameOfContactAlreadyExistInContactsException.java
@@ -1,9 +1,0 @@
-package com.archosResearch.jCHEKS.engine.model;
-
-/**
- *
- * @author Michael Roussel <rousselm4@gmail.com>
- */
-public class NameOfContactAlreadyExistInContactsException extends Exception {
-
-}
