+++ src/com/archosResearch/jCHEKS/engine/model/contact/exception/ContactAlreadyExistException.java
@@ -1,0 +1,9 @@
+package com.archosResearch.jCHEKS.engine.model.contact.exception;
+
+/**
+ *
+ * @author Michael Roussel <rousselm4@gmail.com>
+ */
+public class ContactAlreadyExistException extends Exception {
+
+}
